+++ x/staking/types/keys.go
@@ -30,7 +30,7 @@ const (
 var (
 	// Keys for store prefixes
 	// Last* values are constant during a block.
-	LastValidatorPowerKey = []byte{0x11}              // prefix for each key to a validator index, for bonded validators
+	LastValidatorPowerKey = collections.NewPrefix(17) // prefix for each key to a validator index, for bonded validators
 	LastTotalPowerKey     = collections.NewPrefix(18) // prefix for the total power
 
 	ValidatorsKey             = collections.NewPrefix(33) // prefix for each key to a validator
@@ -124,11 +124,6 @@ func GetValidatorsByPowerIndexKey(validator Validator, powerReduction math.Int, 
 	copy(key[powerBytesLen+2:], operAddrInvr)
 
 	return key
-}
-
-// GetLastValidatorPowerKey creates the bonded validator index key for an operator address
-func GetLastValidatorPowerKey(operator sdk.ValAddress) []byte {
-	return append(LastValidatorPowerKey, address.MustLengthPrefix(operator)...)
 }
 
 // ParseValidatorPowerRankKey parses the validators operator address from power rank key